un pullastre

+++ insignies.php
@@ -55,15 +55,19 @@ include("bbdd.php");
                 <td>
                   <div class="uk-margin-small">
                     <div class="uk-button-group">
+                      <?php if ($_SESSION['rol'] > 1) { ?>
                       <a class="uk-button uk-button-primary uk-button-small" href="insignia.php?<?php echo http_build_query(array(
                                                                                                       'id' => $fila['id']
                                                                                                     )) ?>">Edit</a>
+                                                                                                    <?php }
+                  if ($_SESSION['rol'] > 2) { ?>
                       <a class="uk-button uk-button-danger uk-button-small" href="delete.php?<?php echo http_build_query(array(
                                                                                                     'table' => 'insignies',
                                                                                                     'id' => $fila['id'],
                                                                                                     'name' => $fila['nom'],
                                                                                                     'paginaOrigen' => 'insignies.php'
                                                                                                   )) ?>">Delete</a>
+                                                                                                   <?php } ?>
                     </div>
                   </div>
                 </td>
